move horizon data creation to own file
